requirements.txt
fastapi
python-multipart
uvicorn
pillow
numpy
opencv-python --user
scikit-image